Added dwelling time for single fly
The single fly analysis works good. I need to double check that all the calculations are correct but I am confident that they are.
The problem is that the group analysis does not work. Data for only one of the groups are generated. I am currently trying to figure out why