Update single_fly.Rmd
I have added a completely new way of calculating the dwelling time. The method should be a lot more compatible with the position data compared to the other one. For now, I have kept both of them so we can evaluate them together and next to each other. Some of the lines of code in it are a bit complicated. The data is split into punished and unpunished which makes it easier to read.
I have added a lot of comments to hopefully explain the code