Added a progress bar
In addition to the runtime I also added a progress bar. Currently it shows the previous iteration time, estimated finish time, % completed, and how many flies left. This information is added to a barchart. It does add a few milliseconds to the analysis time but it is negligible compared to the entire duration.