Added normalized data
Do we need to do this on a single fly or only for the group? The process would be similar but as a proof of principle I implemented it into the single fly analysis. The code calculates the pretest and subtracts it from the performance index. However, assume we have a pretest of -0.2, and the preference index is 0.85. Normalizing the PI against the pretest would then result in a PI greater than 1? I "solved" this by capping the maximum PI to 1.