Automated script for creating xml
Till now the files had to be pasted by hand. Now the script takes the two output files from the shiming set-up and creats xml-files for the analysis. A line has to be added to the read.xml file due to the row counter is also pasted in and considered as variable so it has to be removed when reading in the data.
The file also writes all the names in one template yml file, so it dont need to be written by hand.