Progress bar for data integrity
I came by a function called TxtProgressbar. It displays the progress of a for loop in the console. I implemented it to track the progress of the data integrity function. Let me know what you think! It also does not require any additional libraries so I removed the magick one in this example.
I changed the progress indicator to display "DTS", because why not.